refactor: extract runtime routing utils
