+++ test/kit.test.ts
@@ -1,11 +1,6 @@
 import { describe, test, expect, vi } from 'vitest'
-import {
-  type ComposableContext,
-  createLocaleRouteNameGetter,
-  createLocalizedRouteByPathResolver,
-  isRouteLocationPathRaw,
-  prefixable
-} from '../src/runtime/utils'
+import { type ComposableContext, isRouteLocationPathRaw, prefixable } from '../src/runtime/utils'
+import { createLocaleRouteNameGetter, createLocalizedRouteByPathResolver } from '../src/runtime/routing/utils'
 import { getRouteBaseName as _getRouteBaseName } from '#i18n-kit/routing'
 import { createMemoryHistory, createRouter, type RouteLocationRaw } from 'vue-router'
 import { type RouteRecordNameGeneric, type Router } from 'vue-router'
